chore: split hooks into several parts
