+++ packages/runtime/src/hooks/useLang.ts
@@ -1,0 +1,6 @@
+import { usePageData } from './usePageData';
+
+export function useLang(): string {
+  const { page } = usePageData();
+  return page.lang || '';
+}
