+++ packages/runtime/src/hooks/useSiteData.ts
@@ -1,0 +1,6 @@
+import { usePageData } from './usePageData';
+
+export function useSiteData() {
+  const { siteData } = usePageData();
+  return siteData;
+}
