+++ packages/runtime/src/hooks/useVersion.ts
@@ -1,0 +1,6 @@
+import { usePageData } from './usePageData';
+
+export function useVersion(): string {
+  const { page } = usePageData();
+  return page.version || '';
+}
