+++ packages/runtime/src/index.ts
@@ -12,17 +12,13 @@ export {
   useSearchParams,
 } from 'react-router-dom';
 export { Content } from './Content';
-export {
-  DataContext,
-  ThemeContext,
-  useDark,
-  useI18n,
-  useLang,
-  usePageData,
-  useVersion,
-  useViewTransition,
-  useWindowSize,
-} from './hooks';
+export { ThemeContext, useDark } from './hooks/useDark';
+export { useI18n } from './hooks/useI18n';
+export { useLang } from './hooks/useLang';
+export { DataContext, usePageData } from './hooks/usePageData';
+export { useVersion } from './hooks/useVersion';
+export { useViewTransition } from './hooks/useViewTransition';
+export { useWindowSize } from './hooks/useWindowSize';
 export { NoSSR } from './NoSSR';
 export {
   normalizeRoutePath,
